% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.JMcmprsk.R
\name{coef.JMcmprsk}
\alias{coef.JMcmprsk}
\title{Coefficients of longitudinal model}
\usage{
\method{coef}{JMcmprsk}(object, ...)
}
\arguments{
\item{object}{The JMcmprsk object returned by either jmo or jmc function.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Return estimates fixed effects with variable names
}
\description{
Joint modelling for longitutal and censored data with competing risks
}
