% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.JMcmprsk.R
\name{print.JMcmprsk}
\alias{print.JMcmprsk}
\title{Print JMcmprsk}
\usage{
\method{print}{JMcmprsk}(x, digits = 4, ...)
}
\arguments{
\item{x}{Object of class 'JMcmprsk'.}

\item{digits}{The desired number of digits after the decimal point. Default is 4.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print contents of JMcmprsk object.
}
\seealso{
\code{\link{jmc}}
}
\author{
Hong Wang
}
