% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.JMcmprsk.R
\name{summary.JMcmprsk}
\alias{summary.JMcmprsk}
\title{Extraction of standard error and 95\% confidence interval of longitudinal/survival sub-model fixed effects}
\usage{
\method{summary}{JMcmprsk}(object, coeff = c("longitudinal", "survival"), digits = 4, ...)
}
\arguments{
\item{object}{The JMcmprsk object returned by either jmo or jmc function.}

\item{coeff}{The coefficients returned by the JMcmprsk object. Results of longitudinal/survival sub-model will be printed out when typing "longitudinal" or "survival".}

\item{digits}{The number of digits to print out.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Return standard errors of parameters with variable names
}
\description{
Joint modelling for longitutal and censored data with competing risks
}
\references{
\itemize{
\item Elashoff, Robert M., Gang Li, and Ning Li. "A joint model for longitudinal measurements and survival data in the presence of multiple failure types." Biometrics 64.3 (2008): 762-771.
}
}
\seealso{
\code{\link{jmc}}
}
