% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.JMcmprsk.R
\name{coef.JMcmprsk}
\alias{coef.JMcmprsk}
\title{Coefficients of longitudinal/survival sub-model}
\usage{
\method{coef}{JMcmprsk}(object, coeff = c("beta", "alpha", "gamma"), ...)
}
\arguments{
\item{object}{The JMcmprsk object returned by either jmo or jmc function.}

\item{coeff}{The coefficients returned by the JMcmprsk object.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Return estimates fixed effects with variable names
}
\description{
Joint modelling for longitutal and censored data with competing risks
}
\seealso{
\code{\link{jmc}}
}
